requirements.txt
# Core dependencies
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.2.0
torch>=1.8.0
albumentations>=1.0.0

# Logging and experiment tracking
wandb>=0.12.0

# For confusion matrix and metrics calculation
scikit-learn>=0.24.2

# Weight Path handling
gdown>=4.2.0